requirements.txt
appdirs==1.4.4
arrow==0.17.0
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0     
Django==3.1.2
django-heroku==0.3.1       
djangorestframework==3.12.1
gunicorn==20.0.4
imutils==0.5.3
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
times==0.7
whitenoise==5.2.0
